+++ examples/nestjs-api-reference-express/src/main.ts
@@ -2,7 +2,7 @@ import { NestFactory } from '@nestjs/core';
 import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
 import { AppModule } from './app.module';
 
-import { apiReference } from '@scalar/nestjs-api-reference';
+import { apiReference } from '@readyapi/nestjs-api-reference';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);