+++ packages/api-client/src/components/ApiClient/AddressBar.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import { CodeMirror } from '@scalar/use-codemirror'
-import { FlowModal, useModal } from '@scalar/use-modal'
+import { CodeMirror } from '@readyapi/use-codemirror'
+import { FlowModal, useModal } from '@readyapi/use-modal'
 import { isMacOS } from '@scalar/use-tooltip'
 import { useMagicKeys, whenever } from '@vueuse/core'
 import { computed, ref } from 'vue'