+++ packages/api-reference/src/components/Content/Operation/PathResponses/ExampleResponse.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import { ScalarCodeBlock } from '@scalar/components'
+import { ScalarCodeBlock } from '@readyapi/components'
 import { getExampleFromSchema, prettyPrintJson } from '@readyapi/oas-utils'
 
 defineProps<{