+++ packages/api-reference/src/components/Content/Operation/TryRequestButton.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { HttpMethod } from '@readyapi/api-client'
-import { ScalarIcon } from '@scalar/components'
+import { ScalarIcon } from '@readyapi/components'
 import { type TransformedOperation } from '@readyapi/oas-utils'
 import { inject } from 'vue'